feat(tenpayv2): 升级公共组件

+++ src/SKIT.FlurlHttpClient.Wechat.TenpayV2/Models/Pay/ExchangeRate/GetPayExchangeRateResponse.cs
@@ -53,6 +53,6 @@ namespace SKIT.FlurlHttpClient.Wechat.TenpayV2.Models
         /// </summary>
         [Newtonsoft.Json.JsonProperty("rate")]
         [System.Text.Json.Serialization.JsonPropertyName("rate")]
-        public double? ExchangeRate { get; set; } = default!;
+        public decimal? ExchangeRate { get; set; } = default!;
     }
 }
